+++ HISTORY.rst
@@ -11,6 +11,7 @@ History
     * BODY-90: Remove leading underscore from node names. (ericof).
     * Same as BODY-90 but also for relationship fields and fields of schema node attributes (rudaporto).
     * Implement security filters for all REST methods improving BaseResource.apply_security (rudaporto).
+    * BaseFactory now append permission from model class using __acl__ class method (rudaporto).
 
 
 1.1.1 (2016-10-04)